requirements.txt
requests
schedule
python-dotenv
pytest
pandas
